fix(listing-skills): refactorisation part III
