+++ src/modules/skills/hooks/use-skills.ts
@@ -1,6 +1,6 @@
 import { useState } from "react";
 import type { Skill } from "../skills.type";
-import { mockedSkills } from "../skills.mock";
+import { mockedSkills } from "../skills.const";
 
 export const useSkills = () => {
   const [skills, setSkills] = useState<Skill[]>(mockedSkills);
