+++ src/modules/skills/skills.const.ts
@@ -1,3 +1,5 @@
+import type { Skill } from "./skills.type";
+
 export const difficultyConfig = {
   Facile: {
     icon: "🟢",
@@ -83,3 +85,205 @@ export const statCards = [
 ];
 
 export const COLORS = ["#fff", "#facc15", "#a78bfa"];
+
+export const mockedSkills: Skill[] = [
+  {
+    id: "1",
+    title: "Apprendre React",
+    description:
+      "Maîtriser les fondamentaux de React et créer des applications modernes",
+    difficulty: "Moyen",
+    duration: 120,
+    status: "in_progress",
+    category: "Développement Web",
+    createdAt: "2024-01-15T10:00:00Z",
+    updatedAt: "2024-01-20T14:30:00Z",
+    progress: 65,
+    questsCount: 8,
+  },
+  {
+    id: "2",
+    title: "Découvrir TypeScript",
+    description: "Apprendre TypeScript pour un développement plus robuste",
+    difficulty: "Facile",
+    duration: 90,
+    status: "not_started",
+    category: "Développement Web",
+    createdAt: "2024-01-18T09:00:00Z",
+    updatedAt: "2024-01-18T09:00:00Z",
+    progress: 0,
+    questsCount: 5,
+  },
+  {
+    id: "3",
+    title: "Créer un projet Vite",
+    description: "Configurer et optimiser un projet avec Vite",
+    difficulty: "Facile",
+    duration: 60,
+    status: "completed",
+    category: "Outils",
+    createdAt: "2024-01-10T11:00:00Z",
+    updatedAt: "2024-01-12T16:45:00Z",
+    progress: 100,
+    questsCount: 3,
+  },
+  {
+    id: "4",
+    title: "Maîtriser les Hooks React",
+    description:
+      "Utiliser efficacement les hooks personnalisés et les hooks natifs",
+    difficulty: "Difficile",
+    duration: 180,
+    status: "draft",
+    category: "Développement Web",
+    createdAt: "2024-01-22T08:00:00Z",
+    updatedAt: "2024-01-22T08:00:00Z",
+    progress: 0,
+    questsCount: 12,
+  },
+  {
+    id: "5",
+    title: "Design System avec Tailwind",
+    description: "Créer un système de design cohérent avec Tailwind CSS",
+    difficulty: "Moyen",
+    duration: 150,
+    status: "in_progress",
+    category: "Design",
+    createdAt: "2024-01-16T13:00:00Z",
+    updatedAt: "2024-01-21T10:15:00Z",
+    progress: 35,
+    questsCount: 7,
+  },
+  {
+    id: "6",
+    title: "API REST avec Express",
+    description: "Développer des APIs RESTful avec Node.js et Express",
+    difficulty: "Moyen",
+    duration: 200,
+    status: "not_started",
+    category: "Backend",
+    createdAt: "2024-01-25T09:00:00Z",
+    updatedAt: "2024-01-25T09:00:00Z",
+    progress: 0,
+    questsCount: 10,
+  },
+  {
+    id: "7",
+    title: "Base de données MongoDB",
+    description: "Apprendre à utiliser MongoDB pour le stockage de données",
+    difficulty: "Facile",
+    duration: 100,
+    status: "completed",
+    category: "Base de données",
+    createdAt: "2024-01-08T14:00:00Z",
+    updatedAt: "2024-01-15T11:30:00Z",
+    progress: 100,
+    questsCount: 6,
+  },
+  {
+    id: "8",
+    title: "Tests unitaires avec Jest",
+    description: "Écrire des tests unitaires efficaces avec Jest",
+    difficulty: "Moyen",
+    duration: 120,
+    status: "in_progress",
+    category: "Tests",
+    createdAt: "2024-01-20T10:00:00Z",
+    updatedAt: "2024-01-23T16:20:00Z",
+    progress: 45,
+    questsCount: 8,
+  },
+  {
+    id: "9",
+    title: "Déploiement avec Docker",
+    description: "Containeriser et déployer des applications avec Docker",
+    difficulty: "Difficile",
+    duration: 180,
+    status: "draft",
+    category: "DevOps",
+    createdAt: "2024-01-28T08:00:00Z",
+    updatedAt: "2024-01-28T08:00:00Z",
+    progress: 0,
+    questsCount: 15,
+  },
+  {
+    id: "10",
+    title: "GraphQL avec Apollo",
+    description: "Implémenter GraphQL avec Apollo Client et Server",
+    difficulty: "Difficile",
+    duration: 240,
+    status: "not_started",
+    category: "API",
+    createdAt: "2024-01-30T11:00:00Z",
+    updatedAt: "2024-01-30T11:00:00Z",
+    progress: 0,
+    questsCount: 12,
+  },
+  {
+    id: "11",
+    title: "Machine Learning avec Python",
+    description:
+      "Introduction au machine learning avec scikit-learn et TensorFlow",
+    difficulty: "Difficile",
+    duration: 300,
+    status: "not_started",
+    category: "IA/ML",
+    createdAt: "2024-02-01T10:00:00Z",
+    updatedAt: "2024-02-01T10:00:00Z",
+    progress: 0,
+    questsCount: 20,
+  },
+  {
+    id: "12",
+    title: "Cybersécurité Web",
+    description:
+      "Sécuriser les applications web contre les vulnérabilités courantes",
+    difficulty: "Moyen",
+    duration: 160,
+    status: "in_progress",
+    category: "Sécurité",
+    createdAt: "2024-01-27T14:00:00Z",
+    updatedAt: "2024-02-02T09:15:00Z",
+    progress: 25,
+    questsCount: 9,
+  },
+  {
+    id: "13",
+    title: "Microservices avec Spring Boot",
+    description: "Développer des microservices avec Spring Boot et Java",
+    difficulty: "Difficile",
+    duration: 280,
+    status: "draft",
+    category: "Backend",
+    createdAt: "2024-02-03T08:00:00Z",
+    updatedAt: "2024-02-03T08:00:00Z",
+    progress: 0,
+    questsCount: 18,
+  },
+  {
+    id: "14",
+    title: "Flutter pour le mobile",
+    description: "Créer des applications mobiles cross-platform avec Flutter",
+    difficulty: "Moyen",
+    duration: 220,
+    status: "not_started",
+    category: "Mobile",
+    createdAt: "2024-02-05T11:00:00Z",
+    updatedAt: "2024-02-05T11:00:00Z",
+    progress: 0,
+    questsCount: 14,
+  },
+  {
+    id: "15",
+    title: "Blockchain et Smart Contracts",
+    description: "Développer des smart contracts avec Solidity et Ethereum",
+    difficulty: "Difficile",
+    duration: 320,
+    status: "draft",
+    category: "Blockchain",
+    createdAt: "2024-02-07T09:00:00Z",
+    updatedAt: "2024-02-07T09:00:00Z",
+    progress: 0,
+    questsCount: 16,
+  },
+];
